refactor: encapsulate existing parser code as pdt healthcert v1
